New Vue Components to manage TOC, Version, Roles
This is a massive update. All the ms files now interact with a new pattern of Vue Components.
VUEPress sidebar and navbar templates are now updated dated with one custom component each.
The TOC title is moved to a component. Versions are now online even with just one version.
There is a new dev path for docs of the docs, /docs/. Roles are now available for sub-categories
of a version with button at the top of the sidebar.