All changes related to Beacons v0.2 Adding graphs and calls to beacon for current Beacon values.
Extended data shown in a slide in panel. This probably needs a discussion as it is using a backend API operations to get the Beacon data. Mayb we do not want the current  value at all.

The Beacon browser may not be needed by May 1st if the Beacon Explorer is ready, if there is going to be one.