Trying make sense of data feed defintions. The whole idea behind data feeds as (beacons, beacon sets, and dAPIs) is confusing.
There could be a beacon ID which is passed as a datafeed_id, and maybe a dAPI ID passed as a datafeed_id. But dAPIs have names as well. Do these point to any dAPI ID, Beacons set ID or beacon ID.

Any way I have spent a lot of time trying to explain this to readers and need help.